pyproject.toml
[tool.poetry]
name = "tex"
version = "0.1.0"
description = ""
authors = ["Volney Casas <volneyrock@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
djangorestframework = "^3.14.0"
requests = "^2.31.0"
pymongo = "^4.4.1"
mongomock = "^4.1.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
